# Improve `tsort` performance