This PR improves the performance of `tsort` on acyclic graphs (the usual kind of input) by ~12% and reduces the number of memory allocations by ~94%, at the cost of ~38% extra memory in total.

I'd totally understand if the extra memory cost of this change is seen as too prohibitive to merge in. 😁 

For cyclic graphs, allocations improved by ~7% but other performance/allocation changes seem inconclusive.

Benchmarks were ran with `./bench.sh`. They assume that [uutils](https://github.com/uutils/coreutils), a competing coreutils implementation in Rust, has been installed on the path with their tsort at `uutsort`. I'd be more than happy to remove or change `./bench.sh` or move it somewhere else later.

For context, the benchmarks show that for the optimized u-root tsort vs uutils tsort, uutils is still ~3x faster on the test acyclic graph and ~230x faster(!) on the test cyclic graph (probably because the cyclic graph has many, many cycles and uutils tsort doesn't try to handle them all).

**Go benchmarks**

```
goos: darwin
goarch: arm64
pkg: github.com/u-root/u-root/cmds/extra/tsort
cpu: Apple M1
                    │ tsort-bench-before.txt │        tsort-bench-after.txt         │
                    │         sec/op         │    sec/op     vs base                │
TsortAcyclicGraph-8             454.2m ±  0%   398.7m ±  0%  -12.23% (p=0.000 n=15)
TsortCyclicGraph-8              807.0m ± 38%   749.7m ± 48%        ~ (p=0.595 n=15)
geomean                         605.4m         546.7m         -9.70%

                    │ tsort-bench-before.txt │         tsort-bench-after.txt         │
                    │          B/op          │     B/op       vs base                │
TsortAcyclicGraph-8             80.71Mi ± 0%   111.52Mi ± 0%  +38.17% (p=0.000 n=15)
TsortCyclicGraph-8              258.4Mi ± 0%    257.9Mi ± 0%        ~ (p=0.367 n=15)
geomean                         144.4Mi         169.6Mi       +17.43%

                    │ tsort-bench-before.txt │        tsort-bench-after.txt        │
                    │       allocs/op        │  allocs/op   vs base                │
TsortAcyclicGraph-8             2104.7k ± 0%   116.3k ± 0%  -94.48% (p=0.000 n=15)
TsortCyclicGraph-8               517.8k ± 0%   479.2k ± 0%   -7.45% (p=0.000 n=15)
geomean                          1.044M        236.1k       -77.39%
```

**[Hyperfine](https://github.com/sharkdp/hyperfine/) benchmarks**

*Acyclic graph*

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `./tsort-before ./some-random-acyclic-graph.txt` | 486.2 ± 6.0 | 480.6 | 497.1 | 3.39 ± 0.06 |
| `./tsort-after ./some-random-acyclic-graph.txt` | 419.9 ± 3.6 | 415.4 | 428.7 | 2.93 ± 0.04 |
| `uutsort ./some-random-acyclic-graph.txt` | 143.4 ± 1.6 | 141.7 | 148.4 | 1.00 |

*Cyclic graph*

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `./tsort-before ./some-random-cyclic-graph.txt` | 887.1 ± 144.0 | 683.9 | 1094.5 | 229.32 ± 38.58 |
| `./tsort-after ./some-random-cyclic-graph.txt` | 902.3 ± 271.4 | 623.0 | 1531.1 | 233.25 ± 70.91 |
| `uutsort ./some-random-cyclic-graph.txt` | 3.9 ± 0.2 | 3.6 | 5.2 | 1.00 |