@jbduncan 2024-11-24
> I think you might want to squash some of these commits? The performance vs memory size seems acceptable to me.

Oh, thanks for reminding me! I'm on it.

Glad to hear re. performance vs memory. :)