@jbduncan 2024-11-24
I'm removing `bench.sh` and `make_graph_fixtures.go` because the former makes assumptions about the presence of `hyperfine` and which commits to benchmark, and the latter is reported by golangci-lint because it redeclares `main()` in the same package.

So here are their contents for future reference.

**bench.sh**

```sh
#!/usr/bin/env bash

set -euo pipefail

git switch --detach "$(git log main..tsort-performance --oneline | tail -1 | awk "{print \$1}")"
go run make_graph_fixtures.go
go build -o ./tsort-before ./cmds/extra/tsort
go test -run=XXX -bench=Tsort -benchmem -count=15 ./cmds/extra/tsort/... | tee tsort-bench-before.txt

git switch tsort-performance
go build -o ./tsort-after ./cmds/extra/tsort
go test -run=XXX -bench=Tsort -benchmem -count=15 ./cmds/extra/tsort/... | tee tsort-bench-after.txt

go run golang.org/x/perf/cmd/benchstat@latest tsort-bench-before.txt tsort-bench-after.txt | tee tsort-bench-comparison.txt

hyperfine --warmup 5 --shell=none --export-markdown=acyclic.md \
    "./tsort-before ./some-random-acyclic-graph.txt" \
    "./tsort-after ./some-random-acyclic-graph.txt" \
    "uutsort ./some-random-acyclic-graph.txt"
hyperfine --warmup 5 --shell=none --export-markdown=cyclic.md --ignore-failure \
    "./tsort-before ./some-random-cyclic-graph.txt" \
    "./tsort-after ./some-random-cyclic-graph.txt" \
    "uutsort ./some-random-cyclic-graph.txt"
```

**make_graph_fixtures.go**

```go
package main

import (
	"fmt"
	"math/rand"
	"os"
)

func main() {
	func() {
		f, err := os.OpenFile(
			"some-random-acyclic-graph.txt",
			os.O_WRONLY|os.O_CREATE|os.O_TRUNC,
			0o666)
		if err != nil {
			panic(err)
		}

		rnd := rand.New(rand.NewSource(1))
		n := 10_000
		for range 100 * n {
			x := rnd.Intn(n + 1)
			y := rnd.Intn(n + 1)
			_, _ = fmt.Fprintln(f, min(x, y), max(x, y))
		}
	}()

	func() {
		f, err := os.OpenFile(
			"some-random-cyclic-graph.txt",
			os.O_WRONLY|os.O_CREATE|os.O_TRUNC,
			0o666)
		if err != nil {
			panic(err)
		}
		rnd := rand.New(rand.NewSource(1))
		n := 200
		for range 100 * n {
			x := rnd.Intn(n + 1)
			y := rnd.Intn(n + 1)
			_, _ = fmt.Fprintln(f, x, y)
		}
	}()
}
```