@rminnich 2024-11-26
you can put that .go you removed  back and use go generate I think?
https://blog.logrocket.com/using-go-generate-reduce-boilerplate-code/