@jbduncan 2024-11-27
@rminnich Oh, I think you're right, that could work.

That being said, `make_graph_fixtures.go` only existed to make `bench.sh` work, and it writes out the same data that the Go benchmarks in `tsort_test.go` use...

Perhaps I can make `bench.sh` work for any two input commits, but those commits would have to be ones where the Go benchmarks exist, so is it worth it? Not sure. 🤔

Thoughts?