@rminnich 2024-11-27
do as few special things as possible. We've been burned by that. So, would be nice to avoid hyperfine and uutils and make go benchmarking work.