@rminnich 2024-11-27
I saw a pretty good talk at one of the gopher cons about improving this sort of thing. It got down to removing a method call and replacing it with a simple usage of go generics. 

No idea if that is helpful. I like to see this kind of work, looking for performance issues, so this is really nice to see. Also, providing an understanding of where this tsort might be slower than uutils because it's doing a bit more -- we would not want to make it do less just to make a measurement better. 

now that I see what you are doing with hyperfine and uutils, i.e. keeping us informed about where we are slower, I'm inclined to say "let's keep that in there"