@jbduncan 2024-11-28
@rminnich Done! It's not perfect, but I've managed to reintroduce `bench.sh`, clean it up and put it in the `tsort` directory so it doesn't pollute the top level of the project.

Now it expects two input Git commits to benchmark and compare. Also it runs some warmup Go benchmarks to make things run more consistently on my Linux AMD64 laptop.

So I'm ready for another review whenever you are.

(I'm not too sure how we'll deal with your intermediate merge commit in-between my two commits, but I think GitHub's rebase/squash feature will help. :smile:)